requirements.txt
flask
flask-cors
langchain-google-genai
python-docx
python-dotenv